+++ api/models/rules.py
@@ -1,4 +1,5 @@
-from typing import List
+from typing import List, Optional
+from fastapi_camelcase import CamelModel
 from pydantic import BaseModel
 
 
@@ -8,9 +9,21 @@ class InputParameter(BaseModel):
     payload: dict
 
 
-class Rule(BaseModel):
-    """A data generalization rule"""
+class RuleOutput(CamelModel):
+    """A data generalization rule output"""
+    id: str
     name: str
     description: str
-    targetResourceType: str
-    inputParameters: List[InputParameter]
+    resource_type: str
+    input_parameters: List[InputParameter]
+
+
+class RuleInput(BaseModel):
+    """A data generalization rule input"""
+    id: str
+
+
+class RulesBody(CamelModel):
+    """Request body for rules"""
+    resource_types: Optional[List[str]] = None
+    input_filters: Optional[List] = None